[Refactor][Unit test] Refactor dashboard. Add MyMgmtList unit test
