+++ src/components/personalPage/myShareList/CalendarPopup.jsx
@@ -22,13 +22,16 @@ const ClendarPopup = ({ showCalender, closeCalendar }) => {
   const closeInfo = () => setShowInfo(false);
   const now = new Date();
 
+  const openAlertWithMessage = (msg) => {
+    setAlertMessage(msg);
+    openInfo();
+  };
+
   const handleCheckDateTimeRange = () => {
     if (fromToDateTime && fromToDateTime[1] < fromToDateTime[0]) {
-      setAlertMessage('結束時間小於開始時間，請重新選擇結束時間');
-      openInfo();
+      openAlertWithMessage('結束時間小於開始時間，請重新選擇結束時間');
     } else if (fromToDateTime && fromToDateTime[0] <= now) {
-      setAlertMessage('開始時間小於現在時間，請重新選擇開始時間');
-      openInfo();
+      openAlertWithMessage('開始時間小於現在時間，請重新選擇開始時間');
     } else {
       closeCalendar();
     }
